무결성 체크 및 에러 메시지 출력

+++ src/pages/activity.ts
@@ -7,7 +7,7 @@ const router = express.Router();
 router.post("/", (req, res) => {
     const key = req.body.key;
 
-    if (keyCheck(key)) {
+    if (key && keyCheck(key)) {
         db.query("select * from activity", (err, row) => {
             if (err) res.status(500).send("about db error");
             else res.status(200).send({ length: row.length, data: row });
@@ -28,7 +28,7 @@ router.post("/change", (req, res) => {
     const takeaway = req.body.takeaway;
     const image = req.body.image;
 
-    if (keyCheck(key)) {
+    if (key && keyCheck(key)) {
         db.query(
             `UPDATE activity SET name = '${name}', start = '${start}', end = '${end}', introduce = '${introduce}', position = '${position}', takeaway = '${takeaway}', image = '${image}' WHERE id = ${id} and isbackup = 0`,
             (err, row) => {
@@ -52,7 +52,7 @@ router.delete("/:id", (req, res) => {
     const key = req.body.key;
     const id = req.params.id;
 
-    if (keyCheck(key)) {
+    if (key && keyCheck(key)) {
         db.query(
             `DELETE FROM activity WHERE id = ${id} and isbackup = 0`,
             (err, row) => {
